change5

+++ converttocomic.py
@@ -9,6 +9,7 @@ import generatesubtitle
 def convertVideoToComic(youtube_link):
     res = youtubesrt.genSubfromYoutube(youtube_link) #generates subtitle
     if(res==0):
+        print("Generating subs...")
         downloadvideo.DownloadAudio(youtube_link)
         generatesubtitle.transcribe_audio('videoaudio.mp3')
 
@@ -37,8 +38,9 @@ def convertVideoToComic(youtube_link):
         extractframes.extract_frame(video_file,startstr,endstr,count)
     video.release()
     helpers.converttopdf('frames') #convert all images to pdf
-    # helpers.deletefiles('frames')
+    helpers.deletefiles('frames')
     os.remove('videoclip.mp4')
     os.remove('videosub.srt')
-    os.remove('videoaudio.mp3')
+    if(res==0):
+        os.remove('videoaudio.mp3')
